example run with sbatch
