+++ toolbuilder/main.py
@@ -1,4 +1,4 @@
-import argparse, sys, ast, requests
+import argparse, sys, os, subprocess
 
 from toolbuilder.parse_help import from_container
 from toolbuilder.templates import ToolTemplateType
@@ -10,8 +10,11 @@ from toolbuilder.runtest.runner import (
     NotificationOption,
     cli_logging,
     send_slack_notification,
+    execute,
+    add_runtest_args,
 )
-from janis_assistant.engines.enginetypes import EngineType
+
+from janis_assistant.management.configuration import JanisConfiguration
 
 
 def process_args():
@@ -122,75 +125,24 @@ def add_container_args(parser):
 
 
 def do_runtest(args):
-    output = None
-    if args.output:
-        output = ast.literal_eval(args.output)
+    config = None
+    if args.config:
+        config = JanisConfiguration.initial_configuration(path=args.config)
 
-    if args.test_case:
-        test_cases = [args.test_case]
-    else:
-        test_cases = find_test_cases(args.tool)
+    from toolbuilder.runtest import runner
 
-    for tc_name in test_cases:
-        result = run_test_case(
-            tool_id=args.tool,
-            test_case=tc_name,
-            engine=args.engine,
-            output=output,
-            config=args.config,
-        )
-        cli_logging(tc_name, result)
+    runner_path = runner.__file__
 
-    # Send notification to Slack
-    if args.slack_notification_url:
-        option = NotificationOption(
-            url=args.slack_notification_url,
-            tool_name=args.tool,
-            test_case=args.test_case,
-            test_id=args.test_id,
-        )
-        send_slack_notification(result=result, option=option)
+    cli_args = sys.argv[2:]
+    run_test_commands = ["python", runner_path] + cli_args
 
-    # send output to test framework API
-    if args.test_manager_url and args.test_manager_token:
-        option = UpdateStatusOption(
-            url=args.test_manager_url, token=args.test_manager_token
-        )
-        update_status(result, option)
+    precommands = []
+    if config and config.template.id == "spartan":
+        precommands = ["sbatch", "-p", "snowy", "--wrap"]
 
+    commands = precommands + run_test_commands
 
-def add_runtest_args(parser):
-    parser.add_argument("tool", help="Name of tool to test")
-
-    parser.add_argument(
-        "--test-case", help="Name of test case as listed in tool.tests()"
-    )
-
-    parser.add_argument("-e", "--engine", help="engine", default=EngineType.cromwell)
-
-    parser.add_argument("-c", "--config", help="Path to janis config")
-
-    parser.add_argument(
-        "-o", "--output", help="Dry run test by providing a dictionary of output"
-    )
-
-    # For updating test-framework API endpoint
-    parser.add_argument(
-        "--test-manager-url", help="API endpoint to update run status in Test Manager"
-    )
-
-    parser.add_argument(
-        "--test-manager-token", help="Authentication token for Test Manager API"
-    )
-
-    parser.add_argument(
-        "--test-id",
-        help="Test identification to be attached to the notification message",
-    )
-
-    parser.add_argument(
-        "--slack-notification-url", help="Slack webhook to send notifications to"
-    )
+    subprocess.run(commands)
 
 
 if __name__ == "__main__":
